Move code to source/bigquery
