+++ package.json
@@ -31,6 +31,7 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0",
     "@mapbox/vector-tile": "^1.3.0",
+    "base64-arraybuffer": "^0.2.0",
     "cartocolor": "^4.0.0",
     "earcut": "^2.1.2",
     "gl-matrix": "^2.8.1",
