+++ src/sources/MVTWorkers.worker.js
@@ -1,5 +1,4 @@
 import { MVTWorker } from './MVTWorker';
-import { BQMVTWorker } from './BQMVTWorker';
 import { WindshaftWorker } from '../client/WindshaftWorker';
 
 // This file publishes a Web Worker onmessage function that will redirect incoming messages
@@ -8,7 +7,6 @@ import { WindshaftWorker } from '../client/WindshaftWorker';
 
 const workers = {
     MVT: new MVTWorker(),
-    BQMVT: new BQMVTWorker(),
     windshaft: new WindshaftWorker()
 };
